Fix equality check
